part2/functions/diagnoser-agent/requirements.txt
google-cloud-pubsub
google-cloud-secret-manager
requests
openai
anthropic